refactor(Friends): Change friends architeture and files
