+++ ui/src/components/friends/mod.rs
@@ -1,2 +1,6 @@
 pub mod add;
 pub mod friend;
+pub mod friends_list;
+pub mod incoming_requests;
+pub mod outgoing_requests;
+pub mod blocked;
